Changes due to the GUI
Remove the files that became redundant due to the update to a GUI, rename the file gui-sesmg.py to SESMG_EXE.py .
